requirements.dev.txt
coverage==4.5.1
django-nose==1.4.5
factory_boy==2.10.0
freezegun==0.3.10
mock==2.0.0
model_mommy==1.5.1
parameterized==0.6.1
pydevd==1.1.1
requests==2.18.4
Sphinx==1.7.4
sphinx_rtd_theme==0.3.1